Add option for MP airmove/friction/walkmove
g_allowBunnyHopping 1 = No velocity halving
2 = 1 + fixed upmove
3 = 2 + MP airmove
4 = 3 + MP walkmove/friction

The walkmove/friction is nearly identical to MP afaik already, but it seems MP has a tiny bit more friction in practice even though the pm_friction constants are same for both.
I think the main reason for the belief that SP has slidey physics is because the animations play differently and it looks like the player is sliding isntead of slowing their walk.